Return english translation from search service
In partial fulfillment of requirements for this exercise, this commit adds
implementation that requests (i.e search) and returns results from the search.

Tests have been included that demonstrate this, with web mock stubbing the requests.

This implementation fetches all the results from the search.

This will become usefule further down the line.